refactor: simplify chunking code (PROOF-923) (#205)
* rework split

* refactor split

* refactor split

* refactor multiexponentiation

* refactor split code

* drop cruft

* refactor splitting code

* refactor splitting code

* refactor split code

* refactor chunking code

* refactor split

* refactor splitting code

* drop dead code

* refactor

* drop dead code

* refactor

* refactor

* refactor

* drop dead code

* drop dead code

* refactor

* drop dead code

* drop dead code

* refactor

* drop dead code

* drop dead code

* reformat

+++ sxt/base/device/property.cc
@@ -25,14 +25,14 @@ namespace sxt::basdv {
 //--------------------------------------------------------------------------------------------------
 // get_num_devices
 //--------------------------------------------------------------------------------------------------
-int get_num_devices() noexcept {
+unsigned get_num_devices() noexcept {
   static int num_devices = []() noexcept {
     int res;
     auto rcode = cudaGetDeviceCount(&res);
     if (rcode != cudaSuccess) {
-      return 0;
+      return 0u;
     }
-    return res;
+    return static_cast<unsigned>(res);
   }();
   return num_devices;
 }
